Reason: cleanup things

+++ handle_scaling.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
+set -x
+set -e
+
 ad=`jq -r '.ad' config.json`
 fa=`jq -r '.fa' config.json`
 md=`jq -r '.md' config.json`
@@ -23,14 +26,16 @@ echo "input diffusion metrics set"
 #### loop through metrics and generate stats text files ####
 for MET in ${METRIC}
 do
-	metric=$(eval "echo \$${MET}")
+	if [ ! -f ${MET}.nii.gz ]; then
+		metric=$(eval "echo \$${MET}")
 
-	# handle scaling issues
-	median_val=$(eval "fslstats ${metric} -P 50")
-	if [[ $median_val < 0.01 ]]; then 
-		fslmaths ${metric} -mul 1000 ./${MET}.nii.gz
-		metric="./${MET}.nii.gz"
-	else
-		cp ${metric} ./
+		# handle scaling issues
+		median_val=$(eval "fslstats ${metric} -P 50")
+		if [[ $median_val < 0.01 ]]; then 
+			fslmaths ${metric} -mul 1000 ./${MET}.nii.gz
+			metric="./${MET}.nii.gz"
+		else
+			cp ${metric} ./
+		fi
 	fi
 done
